+++ changelog.md
@@ -1,6 +1,12 @@
 
 
 # Changelog
+* 2019-12-08
+  * Fixed a css scale bug in the viewport() function.
+* 2019-12-07
+  * Added a fallback CSS raster (inline PNG).
+  * Added the drawConfig for lines, polygons, ellipse, triangles, bezier curves and image control lines.
+  * Added a new demo for bezier spline interpolation.
 * 2019-12-04 (1.6.0)
   * Added relative positioned zooming.
   * Added offsetX and offsetY params.